Moved Explorer data to presentation layer
